+++ Source/Application/VQEngine_Render.cpp
@@ -488,10 +488,10 @@ void VQEngine::RenderThread_PreRender()
 
 void VQEngine::RenderThread_Render()
 {
-	if (mbRenderEnvironmentMapIrradiance.load())
+	if (mbEnvironmentMapPreFilter.load())
 	{
-		RenderEnvironmentMapCubeFaces(mResources_MainWnd.EnvironmentMap);
-		mbRenderEnvironmentMapIrradiance.store(false);
+		PreFilterEnvironmentMap(mResources_MainWnd.EnvironmentMap);
+		mbEnvironmentMapPreFilter.store(false);
 	}
 
 	RenderThread_RenderMainWindow();
@@ -629,309 +629,6 @@ void VQEngine::RenderThread_RenderDebugWindow()
 #endif
 }
 
-void VQEngine::RenderEnvironmentMapCubeFaces(FEnvironmentMap& env)
-{
-	using namespace DirectX;
-
-	FWindowRenderContext& ctx = mRenderer.GetWindowRenderContext(mpWinMain->GetHWND());
-	ID3D12GraphicsCommandList*& pCmd = ctx.pCmdList_GFX;
-
-	SCOPED_GPU_MARKER(pCmd, "RenderEnvironmentMapCubeFaces");
-
-	constexpr int MIP_LEVELS = 12;
-	constexpr int NUM_CUBE_FACES = 6;
-	
-
-	const SRV& srvEnv = mRenderer.GetSRV(env.SRV_HDREnvironment);
-	const SRV& srvEnvDown = mRenderer.GetSRV(env.SRV_HDREnvironmentDownsampled);
-	const SRV& srvIrrDiffuse   = mRenderer.GetSRV(env.SRV_IrradianceDiff);
-	const SRV& srvIrrSpcecular = mRenderer.GetSRV(env.SRV_IrradianceSpec);
-
-	const XMFLOAT4X4 f16proj = MakePerspectiveProjectionMatrix(PI_DIV2, 1.0f, 0.1f, 10.0f);
-	const XMMATRIX proj = XMLoadFloat4x4(&f16proj);
-	
-	struct cb0_t { XMMATRIX viewProj[NUM_CUBE_FACES]; };
-	struct cb1_t { float Roughness; float ViewDimX; float ViewDimY; int MIP; };
-
-	// Downsample
-	{
-		SCOPED_GPU_MARKER(pCmd, "EnvironmentMapDownsample");
-		
-		pCmd->SetPipelineState(mRenderer.GetPSO(HDR_FP16_SWAPCHAIN_PSO));
-		pCmd->SetGraphicsRootSignature(mRenderer.GetRootSignature(1));
-
-		pCmd->SetGraphicsRootDescriptorTable(0, srvEnv.GetGPUDescHandle());
-		
-
-		// set viewport & render target
-		int HDREnvironmentSizeX = 0; int HDREnvironmentSizeY = 0;
-		mRenderer.GetTextureDimensions(env.Tex_HDREnvironment, HDREnvironmentSizeX, HDREnvironmentSizeY);
-		assert(HDREnvironmentSizeX != 0 && HDREnvironmentSizeY != 0);
-
-		const float           RenderResolutionX = static_cast<float>(HDREnvironmentSizeX >> 3);
-		const float           RenderResolutionY = static_cast<float>(HDREnvironmentSizeY >> 3);
-		D3D12_VIEWPORT        viewport{ 0.0f, 0.0f, RenderResolutionX, RenderResolutionY, 0.0f, 1.0f };
-		D3D12_RECT            scissorsRect{ 0, 0, (LONG)RenderResolutionX, (LONG)RenderResolutionY };
-		const RTV& rtv = mRenderer.GetRTV(env.RTV_HDREnvironmentDownsampled);
-		pCmd->RSSetViewports(1, &viewport);
-		pCmd->RSSetScissorRects(1, &scissorsRect);
-		pCmd->OMSetRenderTargets(1, &rtv.GetCPUDescHandle(), false, NULL);
-
-		// draw fullscreen triangle
-		const auto      VBIBIDs = mBuiltinMeshes[EBuiltInMeshes::TRIANGLE].GetIABufferIDs();
-		const BufferID& IB_ID   = VBIBIDs.second;
-		const IBV&      ib      = mRenderer.GetIndexBufferView(IB_ID);
-		pCmd->IASetPrimitiveTopology(D3D_PRIMITIVE_TOPOLOGY_TRIANGLELIST);
-		pCmd->IASetVertexBuffers(0, 1, NULL);
-		pCmd->IASetIndexBuffer(&ib);
-		pCmd->DrawIndexedInstanced(3, 1, 0, 0, 0);
-
-		pCmd->ResourceBarrier(1, &CD3DX12_RESOURCE_BARRIER::Transition(
-			  mRenderer.GetTextureResource(env.Tex_HDREnvironmentDownsampled)
-			, D3D12_RESOURCE_STATE_RENDER_TARGET
-			, D3D12_RESOURCE_STATE_PIXEL_SHADER_RESOURCE)
-		); // Transition resource for read
-	}
-
-	// Diffuse Irradiance Convolution
-	{
-		SCOPED_GPU_MARKER(pCmd, "DiffuseIrradianceCubemap");
-
-		// Since calculating the diffuse irradiance integral takes a long time,
-		// we opt-in for drawing cube faces separately instead of doing all at one 
-		// go using GS for RT slicing. Otherwise, a TDR can occur when drawing all
-		// 6 faces per draw call as the integration takes long for each face.
-		constexpr bool DRAW_CUBE_FACES_SEPARATELY = true;
-		//------------------------------------------------------------------------
-
-		const RTV& rtv = mRenderer.GetRTV(env.RTV_IrradianceDiff);
-
-		// Viewport & Scissors
-		int w, h, d;
-		mRenderer.GetTextureDimensions(env.Tex_IrradianceDiff, w, h, d);
-		D3D12_VIEWPORT viewport{ 0.0f, 0.0f, static_cast<float>(w), static_cast<float>(h), 0.0f, 1.0f };
-		D3D12_RECT     scissorsRect{ 0, 0, (LONG)w, (LONG)h };
-		pCmd->RSSetViewports(1, &viewport);
-		pCmd->RSSetScissorRects(1, &scissorsRect);
-
-		// geometry input
-		const Mesh& mesh = mBuiltinMeshes[EBuiltInMeshes::CUBE];
-		const auto VBIBIDs = mesh.GetIABufferIDs();
-		const uint32 NumIndices = mesh.GetNumIndices();
-		const BufferID& VB_ID = VBIBIDs.first;
-		const BufferID& IB_ID = VBIBIDs.second;
-		const VBV& vb = mRenderer.GetVertexBufferView(VB_ID);
-		const IBV& ib = mRenderer.GetIndexBufferView(IB_ID);
-
-		pCmd->IASetPrimitiveTopology(D3D_PRIMITIVE_TOPOLOGY_TRIANGLELIST);
-		pCmd->IASetVertexBuffers(0, 1, &vb);
-		pCmd->IASetIndexBuffer(&ib);
-
-		if constexpr (DRAW_CUBE_FACES_SEPARATELY)
-		{
-			pCmd->SetPipelineState(mRenderer.GetPSO(CUBEMAP_CONVOLUTION_DIFFUSE_PER_FACE_PSO));
-			pCmd->SetGraphicsRootSignature(mRenderer.GetRootSignature(10));
-			pCmd->SetGraphicsRootDescriptorTable(2, srvEnvDown.GetGPUDescHandle());
-			pCmd->SetGraphicsRootDescriptorTable(3, srvEnv.GetGPUDescHandle());
-
-			for (int face = 0; face < NUM_CUBE_FACES; ++face)
-			{
-				D3D12_CPU_DESCRIPTOR_HANDLE rtvHandle = rtv.GetCPUDescHandle(face);
-				
-				D3D12_GPU_VIRTUAL_ADDRESS cbAddr0 = {};
-				D3D12_GPU_VIRTUAL_ADDRESS cbAddr1 = {};
-				cb0_t* pCB0 = {};
-				cb1_t* pCB1 = {};
-				ctx.mDynamicHeap_ConstantBuffer.AllocConstantBuffer(sizeof(cb0_t), (void**)(&pCB0), &cbAddr0);
-				ctx.mDynamicHeap_ConstantBuffer.AllocConstantBuffer(sizeof(cb1_t), (void**)(&pCB1), &cbAddr1);
-				pCB1->ViewDimX = static_cast<float>(w);
-				pCB1->ViewDimY = static_cast<float>(h);
-				pCB1->Roughness = 0.0f;
-				pCB0->viewProj[0] = Texture::CubemapUtility::CalculateViewMatrix(face) * proj;
-
-				pCmd->SetGraphicsRootConstantBufferView(1, cbAddr1);
-				pCmd->SetGraphicsRootConstantBufferView(0, cbAddr0);
-				pCmd->OMSetRenderTargets(1, &rtvHandle, TRUE, NULL);
-				pCmd->DrawIndexedInstanced(NumIndices, 1, 0, 0, 0);
-			}
-		}
-
-		else // Draw Instanced Cubes (1x instance / cube face) w/ GS RT slicing
-		{
-			D3D12_CPU_DESCRIPTOR_HANDLE rtvHandle = rtv.GetCPUDescHandle();
-
-			D3D12_GPU_VIRTUAL_ADDRESS cbAddr0 = {};
-			D3D12_GPU_VIRTUAL_ADDRESS cbAddr1 = {};
-			cb0_t* pCB0 = {};
-			cb1_t* pCB1 = {};
-			ctx.mDynamicHeap_ConstantBuffer.AllocConstantBuffer(sizeof(cb0_t), (void**)(&pCB0), &cbAddr0);
-			ctx.mDynamicHeap_ConstantBuffer.AllocConstantBuffer(sizeof(cb1_t), (void**)(&pCB1), &cbAddr1);
-			pCB1->ViewDimX = static_cast<float>(w);
-			pCB1->ViewDimY = static_cast<float>(h);
-			pCB1->MIP = 0;
-			pCB1->Roughness = 0.0f;
-			for (int face = 0; face < NUM_CUBE_FACES; ++face)
-			{
-				pCB0->viewProj[face] = Texture::CubemapUtility::CalculateViewMatrix(face) * proj;
-			}
-
-			pCmd->SetPipelineState(mRenderer.GetPSO(CUBEMAP_CONVOLUTION_DIFFUSE_PSO));
-			pCmd->SetGraphicsRootSignature(mRenderer.GetRootSignature(10));
-			pCmd->SetGraphicsRootDescriptorTable(2, srvEnvDown.GetGPUDescHandle());
-			pCmd->SetGraphicsRootDescriptorTable(3, srvEnv.GetGPUDescHandle()); // ?
-			pCmd->SetGraphicsRootConstantBufferView(0, cbAddr0);
-			pCmd->SetGraphicsRootConstantBufferView(1, cbAddr1);
-			pCmd->OMSetRenderTargets(1, &rtvHandle, TRUE, NULL);
-			pCmd->DrawIndexedInstanced(NumIndices, NUM_CUBE_FACES, 0, 0, 0);
-		}
-
-		const CD3DX12_RESOURCE_BARRIER pBarriers[] =
-		{
-			  CD3DX12_RESOURCE_BARRIER::Transition(mRenderer.GetTextureResource(env.Tex_IrradianceDiff), D3D12_RESOURCE_STATE_RENDER_TARGET, D3D12_RESOURCE_STATE_NON_PIXEL_SHADER_RESOURCE)
-		};
-		pCmd->ResourceBarrier(_countof(pBarriers), pBarriers);
-	}
-	
-	// Blur Diffuse Irradiance
-	for (int face = 0; face < NUM_CUBE_FACES; ++face)
-	{
-		std::string marker = "Blur Face[";
-		marker += std::to_string(face);
-		marker += "]";
-		SCOPED_GPU_MARKER(pCmd, marker.c_str());
-
-
-		// compute dispatch dimensions
-		int InputImageWidth = 0;
-		int InputImageHeight = 0;
-		int InputImageNumSlices = 0;
-		mRenderer.GetTextureDimensions(env.Tex_IrradianceDiff, InputImageWidth, InputImageHeight, InputImageNumSlices);
-		assert(InputImageNumSlices == NUM_CUBE_FACES);
-
-		const SRV& srv = mRenderer.GetSRV(env.SRV_IrradianceDiffFaces[face]);
-
-		constexpr int DispatchGroupDimensionX = 8;
-		constexpr int DispatchGroupDimensionY = 8;
-		const     int DispatchX = (InputImageWidth  + (DispatchGroupDimensionX - 1)) / DispatchGroupDimensionX;
-		const     int DispatchY = (InputImageHeight + (DispatchGroupDimensionY - 1)) / DispatchGroupDimensionY;
-		constexpr int DispatchZ = 1;
-
-		const UAV& uav_BlurIntermediate = mRenderer.GetUAV(env.UAV_BlurTemp);
-		const UAV& uav_BlurOutput       = mRenderer.GetUAV(env.UAV_IrradianceDiffBlurred);
-		const SRV& srv_BlurIntermediate = mRenderer.GetSRV(env.SRV_BlurTemp);
-		ID3D12Resource* pRscBlurIntermediate = mRenderer.GetTextureResource(env.Tex_BlurTemp);
-		ID3D12Resource* pRscBlurOutput       = mRenderer.GetTextureResource(env.Tex_IrradianceDiffBlurred);
-
-		struct FBlurParams { int iImageSizeX; int iImageSizeY; };
-		FBlurParams* pBlurParams = nullptr;
-
-		D3D12_GPU_VIRTUAL_ADDRESS cbAddr = {};
-		ctx.mDynamicHeap_ConstantBuffer.AllocConstantBuffer(sizeof(FBlurParams), (void**)&pBlurParams, &cbAddr);
-		pBlurParams->iImageSizeX = InputImageWidth;
-		pBlurParams->iImageSizeY = InputImageHeight;
-
-		{
-			SCOPED_GPU_MARKER(pCmd, "BlurX");
-			pCmd->SetPipelineState(mRenderer.GetPSO(EBuiltinPSOs::GAUSSIAN_BLUR_CS_NAIVE_X_PSO));
-			pCmd->SetComputeRootSignature(mRenderer.GetRootSignature(11));
-
-			pCmd->SetComputeRootDescriptorTable(0, srv.GetGPUDescHandle());
-			pCmd->SetComputeRootDescriptorTable(1, uav_BlurIntermediate.GetGPUDescHandle());
-			pCmd->SetComputeRootConstantBufferView(2, cbAddr);
-			pCmd->Dispatch(DispatchX, DispatchY, DispatchZ);
-
-			const CD3DX12_RESOURCE_BARRIER pBarriers[] =
-			{
-				  CD3DX12_RESOURCE_BARRIER::Transition(pRscBlurIntermediate, D3D12_RESOURCE_STATE_UNORDERED_ACCESS, D3D12_RESOURCE_STATE_NON_PIXEL_SHADER_RESOURCE)
-			};
-			pCmd->ResourceBarrier(_countof(pBarriers), pBarriers);
-		}
-		{
-			SCOPED_GPU_MARKER(pCmd, "BlurY");
-			pCmd->SetPipelineState(mRenderer.GetPSO(EBuiltinPSOs::GAUSSIAN_BLUR_CS_NAIVE_Y_PSO));
-			pCmd->SetComputeRootDescriptorTable(0, srv_BlurIntermediate.GetGPUDescHandle());
-			pCmd->SetComputeRootDescriptorTable(1, uav_BlurOutput.GetGPUDescHandle(face));
-			pCmd->SetComputeRootConstantBufferView(2, cbAddr);
-			pCmd->Dispatch(DispatchX, DispatchY, DispatchZ);
-
-			const CD3DX12_RESOURCE_BARRIER pBarriers[] =
-			{
-				CD3DX12_RESOURCE_BARRIER::Transition(pRscBlurIntermediate, D3D12_RESOURCE_STATE_NON_PIXEL_SHADER_RESOURCE, D3D12_RESOURCE_STATE_UNORDERED_ACCESS)
-			};
-			pCmd->ResourceBarrier(_countof(pBarriers), pBarriers);
-		}
-	} // for_each face
-	
-	// transition blurred diffuse irradiance map resource
-	{
-		const CD3DX12_RESOURCE_BARRIER pBarriers[] =
-		{
-			  CD3DX12_RESOURCE_BARRIER::Transition(mRenderer.GetTextureResource(env.Tex_IrradianceDiffBlurred), D3D12_RESOURCE_STATE_UNORDERED_ACCESS, D3D12_RESOURCE_STATE_PIXEL_SHADER_RESOURCE)
-		};
-		pCmd->ResourceBarrier(_countof(pBarriers), pBarriers);
-	}
-
-	// Specular Irradiance
-	if(false)
-	{
-		pCmd->SetPipelineState(mRenderer.GetPSO(CUBEMAP_CONVOLUTION_SPECULAR_PSO));
-		pCmd->SetGraphicsRootDescriptorTable(3, srvIrrSpcecular.GetGPUDescHandle());
-		SCOPED_GPU_MARKER(pCmd, "SpecularIrradianceCubemap");
-		for (int mip = 0; mip < MIP_LEVELS; ++mip)
-		{
-			D3D12_GPU_VIRTUAL_ADDRESS cbAddr0 = {};
-			D3D12_GPU_VIRTUAL_ADDRESS cbAddr1 = {};
-			cb0_t* pCB0 = {};
-			cb1_t* pCB1 = {};
-			ctx.mDynamicHeap_ConstantBuffer.AllocConstantBuffer(sizeof(cb0_t), (void**)(&pCB0), &cbAddr0);
-			ctx.mDynamicHeap_ConstantBuffer.AllocConstantBuffer(sizeof(cb1_t), (void**)(&pCB1), &cbAddr1);
-			for (int face = 0; face < 6; ++face)
-			{
-				pCB0->viewProj[face] = Texture::CubemapUtility::CalculateViewMatrix(face) * proj;
-			}
-
-			const RTV& rtv = mRenderer.GetRTV(env.RTV_IrradianceSpec);
-			D3D12_CPU_DESCRIPTOR_HANDLE rtvHandle = rtv.GetCPUDescHandle(mip * 6);
-
-			pCB1->Roughness = 0.0f;
-			pCB1->ViewDimX = static_cast<float>(2048 >> mip);
-			pCB1->ViewDimY = static_cast<float>(2048 >> mip);
-			pCB1->MIP = mip;
-
-			pCmd->SetGraphicsRootConstantBufferView(0, cbAddr0);
-			pCmd->SetGraphicsRootConstantBufferView(1, cbAddr1);
-
-			pCmd->OMSetRenderTargets(1, &rtvHandle, TRUE, NULL);
-
-			D3D12_VIEWPORT viewport{ 0.0f, 0.0f, pCB1->ViewDimX, pCB1->ViewDimY, 0.0f, 1.0f };
-			D3D12_RECT     scissorsRect{ 0, 0, (LONG)pCB1->ViewDimX, (LONG)pCB1->ViewDimY };
-			pCmd->RSSetViewports(1, &viewport);
-			pCmd->RSSetScissorRects(1, &scissorsRect);
-
-			const Mesh& mesh = mBuiltinMeshes[EBuiltInMeshes::CUBE];
-			const auto VBIBIDs = mesh.GetIABufferIDs();
-			const uint32 NumIndices = mesh.GetNumIndices();
-			const uint32 NumInstances = 6; // 1 cube / face (cube is used for determining sample vector)
-			const BufferID& VB_ID = VBIBIDs.first;
-			const BufferID& IB_ID = VBIBIDs.second;
-			const VBV& vb = mRenderer.GetVertexBufferView(VB_ID);
-			const IBV& ib = mRenderer.GetIndexBufferView(IB_ID);
-
-			pCmd->IASetPrimitiveTopology(D3D_PRIMITIVE_TOPOLOGY_TRIANGLELIST);
-			pCmd->IASetVertexBuffers(0, 1, &vb);
-			pCmd->IASetIndexBuffer(&ib);
-
-			pCmd->DrawIndexedInstanced(NumIndices, NumInstances, 0, 0, 0);
-		}
-	}
-
-	const CD3DX12_RESOURCE_BARRIER pBarriers[] =
-	{
-		  CD3DX12_RESOURCE_BARRIER::Transition(mRenderer.GetTextureResource(env.Tex_IrradianceDiff), D3D12_RESOURCE_STATE_NON_PIXEL_SHADER_RESOURCE, D3D12_RESOURCE_STATE_PIXEL_SHADER_RESOURCE)
-		, CD3DX12_RESOURCE_BARRIER::Transition(mRenderer.GetTextureResource(env.Tex_IrradianceSpec), D3D12_RESOURCE_STATE_RENDER_TARGET, D3D12_RESOURCE_STATE_PIXEL_SHADER_RESOURCE)
-	};
-	pCmd->ResourceBarrier(_countof(pBarriers), pBarriers);
-}
 
 HRESULT VQEngine::RenderThread_RenderMainWindow_LoadingScreen(FWindowRenderContext& ctx)
 {
@@ -1376,6 +1073,8 @@ void VQEngine::RenderSceneColor(FWindowRenderContext& ctx, const FSceneView& Sce
 	using namespace VQ_SHADER_DATA;
 	// set PerFrame constants
 	{
+		const CBV_SRV_UAV& NullCubemapSRV = mRenderer.GetSRV(mResources_MainWnd.SRV_NullCubemap);
+
 		constexpr UINT PerFrameRSBindSlot = 3;
 		PerFrameData* pPerFrame = {};
 		D3D12_GPU_VIRTUAL_ADDRESS cbAddr = {};
@@ -1403,11 +1102,11 @@ void VQEngine::RenderSceneColor(FWindowRenderContext& ctx, const FSceneView& Sce
 		//pCmd->SetGraphicsRootDescriptorTable(7, mRenderer.GetSRV(mResources_MainWnd.SRV_NullCubemap).GetGPUDescHandle());
 		pCmd->SetGraphicsRootDescriptorTable(7, bDrawEnvironmentMap
 			? mRenderer.GetSRV(mResources_MainWnd.EnvironmentMap.SRV_IrradianceDiffBlurred).GetGPUDescHandle(0)
-			: mRenderer.GetSRV(mResources_MainWnd.SRV_NullCubemap).GetGPUDescHandle()
+			: NullCubemapSRV.GetGPUDescHandle()
 		);
 		pCmd->SetGraphicsRootDescriptorTable(8, bDrawEnvironmentMap
-			? mRenderer.GetSRV(mResources_MainWnd.EnvironmentMap.SRV_IrradianceDiff).GetGPUDescHandle(0)
-			: mRenderer.GetSRV(mResources_MainWnd.SRV_NullCubemap).GetGPUDescHandle()
+			? mRenderer.GetSRV(mResources_MainWnd.EnvironmentMap.SRV_IrradianceSpec).GetGPUDescHandle(0)
+			: NullCubemapSRV.GetGPUDescHandle()
 		);
 	}
