+++ Source/Application/VQEngine_Update.cpp
@@ -412,19 +412,6 @@ void VQEngine::WaitUntilRenderingFinishes()
 	while (mNumRenderLoopsExecuted != mNumUpdateLoopsExecuted);
 }
 
-const FEnvironmentMapDescriptor& VQEngine::GetEnvironmentMapDesc(const std::string& EnvMapName) const
-{
-	static const FEnvironmentMapDescriptor DEFAULT_ENV_MAP_DESC = { "ENV_MAP_NOT_FOUND", "", 0.0f };
-	const bool bEnvMapNotFound = mLookup_EnvironmentMapDescriptors.find(EnvMapName) == mLookup_EnvironmentMapDescriptors.end();
-	if (bEnvMapNotFound)
-	{
-		Log::Error("Environment Map %s not found.", EnvMapName.c_str());
-	}
-	return  bEnvMapNotFound
-		? DEFAULT_ENV_MAP_DESC
-		: mLookup_EnvironmentMapDescriptors.at(EnvMapName);
-}
-
 // ---------------------------------------------------------------------
 
 void VQEngine::UpdateThread_UpdateScene_MainWnd(const float dt)
@@ -488,172 +475,6 @@ void VQEngine::Load_SceneData_Dispatch()
 	if (!SceneRep.EnvironmentMapPreset.empty()) 
 	{ 
 		mWorkers_TextureLoading.AddTask([=]() { LoadEnvironmentMap(SceneRep.EnvironmentMapPreset); });
-	}
-}
-
-void VQEngine::LoadEnvironmentMap(const std::string& EnvMapName)
-{
-	assert(EnvMapName.size() != 0);
-	FEnvironmentMap& env = mResources_MainWnd.EnvironmentMap;
-
-	// if already loaded, unload it
-	if (env.Tex_HDREnvironment != INVALID_ID)
-	{
-		assert(env.SRV_HDREnvironment != INVALID_ID);
-		UnloadEnvironmentMap();
-	}
-
-	const FEnvironmentMapDescriptor& desc = this->GetEnvironmentMapDesc(EnvMapName);
-	std::vector<std::string>::iterator it = std::find(mResourceNames.mEnvironmentMapPresetNames.begin(), mResourceNames.mEnvironmentMapPresetNames.end(), EnvMapName);
-	const size_t ActiveEnvMapIndex = it - mResourceNames.mEnvironmentMapPresetNames.begin();
-	
-	if (desc.FilePath.empty()) // check whether the env map was found or not
-	{
-		Log::Error("Couldn't find Environment Map: %s", EnvMapName.c_str());
-		Log::Warning("Have you run Scripts/DownloadAssets.bat?");
-		return;
-	}
-
-
-	Log::Info("Loading Environment Map: %s", EnvMapName.c_str());
-
-	// HDR map
-	env.Tex_HDREnvironment = mRenderer.CreateTextureFromFile(desc.FilePath.c_str());
-	env.SRV_HDREnvironment = mRenderer.CreateAndInitializeSRV(env.Tex_HDREnvironment);
-	env.MaxContentLightLevel = static_cast<int>(desc.MaxContentLightLevel);
-
-	// HDR Map Downsampled 
-	int HDREnvironmentSizeX = 0;
-	int HDREnvironmentSizeY = 0;
-	mRenderer.GetTextureDimensions(env.Tex_HDREnvironment, HDREnvironmentSizeX, HDREnvironmentSizeY);
-	{
-		TextureCreateDesc tdesc("EnvMap_Downsampled");
-		tdesc.bCubemap = false;
-		tdesc.bGenerateMips = false;
-		tdesc.pData = nullptr;
-		tdesc.d3d12Desc.Width  = HDREnvironmentSizeX >> 3;
-		tdesc.d3d12Desc.Height = HDREnvironmentSizeY >> 3;
-		tdesc.d3d12Desc.Format = DXGI_FORMAT_R16G16B16A16_FLOAT;
-		tdesc.d3d12Desc.Dimension = D3D12_RESOURCE_DIMENSION_TEXTURE2D;
-		tdesc.d3d12Desc.DepthOrArraySize = 1;
-		tdesc.d3d12Desc.MipLevels = 1;
-		tdesc.d3d12Desc.SampleDesc = { 1, 0 };
-		tdesc.d3d12Desc.Flags = D3D12_RESOURCE_FLAG_ALLOW_RENDER_TARGET;
-		tdesc.ResourceState = D3D12_RESOURCE_STATES::D3D12_RESOURCE_STATE_RENDER_TARGET;
-		env.Tex_HDREnvironmentDownsampled = mRenderer.CreateTexture(tdesc);
-		
-		env.SRV_HDREnvironmentDownsampled = mRenderer.CreateAndInitializeSRV(env.Tex_HDREnvironmentDownsampled);
-		env.RTV_HDREnvironmentDownsampled = mRenderer.CreateRTV();
-		mRenderer.InitializeRTV(env.RTV_HDREnvironmentDownsampled, 0, env.Tex_HDREnvironmentDownsampled);
-	}
-
-
-	// Create Irradiance Map Textures
-	TextureCreateDesc tdesc("EnvMap_IrradianceDiff");
-	tdesc.bCubemap = true;
-	tdesc.bGenerateMips = true;
-	tdesc.pData = nullptr;
-	tdesc.d3d12Desc.Height = 32;
-	tdesc.d3d12Desc.Width  = 32;
-	tdesc.d3d12Desc.Format = DXGI_FORMAT_R16G16B16A16_FLOAT;
-	tdesc.d3d12Desc.Dimension = D3D12_RESOURCE_DIMENSION_TEXTURE2D;
-	tdesc.d3d12Desc.DepthOrArraySize = 6;
-	tdesc.d3d12Desc.MipLevels = 1;
-	tdesc.d3d12Desc.SampleDesc = { 1, 0 };
-	tdesc.d3d12Desc.Flags = D3D12_RESOURCE_FLAG_ALLOW_RENDER_TARGET;
-	tdesc.ResourceState = D3D12_RESOURCE_STATES::D3D12_RESOURCE_STATE_RENDER_TARGET;
-	env.Tex_IrradianceDiff = mRenderer.CreateTexture(tdesc);
-
-	tdesc.TexName = "EnvMap_IrradianceDiffBlurred";
-	tdesc.d3d12Desc.Flags = D3D12_RESOURCE_FLAG_ALLOW_UNORDERED_ACCESS;
-	tdesc.ResourceState = D3D12_RESOURCE_STATES::D3D12_RESOURCE_STATE_UNORDERED_ACCESS;
-	env.Tex_IrradianceDiffBlurred = mRenderer.CreateTexture(tdesc);
-
-	tdesc.TexName = "EnvMap_BlurImmediateTemp";
-	tdesc.bCubemap = false;
-	tdesc.d3d12Desc.DepthOrArraySize = 1;
-	env.Tex_BlurTemp = mRenderer.CreateTexture(tdesc);
-
-	tdesc.TexName = "EnvMap_IrradianceSpec";
-	tdesc.d3d12Desc.DepthOrArraySize = 6;
-	tdesc.bCubemap = true;
-	tdesc.d3d12Desc.Flags = D3D12_RESOURCE_FLAG_ALLOW_RENDER_TARGET;
-	tdesc.ResourceState = D3D12_RESOURCE_STATES::D3D12_RESOURCE_STATE_RENDER_TARGET;
-	tdesc.d3d12Desc.MipLevels = Image::CalculateMipLevelCount(tdesc.d3d12Desc.Width, tdesc.d3d12Desc.Height);
-	env.Tex_IrradianceSpec = mRenderer.CreateTexture(tdesc);
-
-	const int& NUM_MIPS = tdesc.d3d12Desc.MipLevels;
-
-	// Create Irradiance Map RTVs & UAVs
-	env.RTV_IrradianceDiff = mRenderer.CreateRTV(6);
-	env.RTV_IrradianceSpec = mRenderer.CreateRTV(6 * NUM_MIPS);
-	env.UAV_IrradianceDiffBlurred = mRenderer.CreateUAV(6);
-	env.UAV_BlurTemp = mRenderer.CreateUAV();
-	for (int face = 0; face < 6; ++face)
-	{
-		constexpr int MIP_LEVEL = 0;
-		mRenderer.InitializeRTV(env.RTV_IrradianceDiff, face, env.Tex_IrradianceDiff, face, MIP_LEVEL);
-		mRenderer.InitializeUAV(env.UAV_IrradianceDiffBlurred, face, env.Tex_IrradianceDiffBlurred);
-	}
-	mRenderer.InitializeUAV(env.UAV_BlurTemp, 0, env.Tex_BlurTemp);
-
-	for (int  mip = 0; mip<NUM_MIPS; ++mip ) 
-	for (int face = 0; face < 6    ; ++face)  
-		mRenderer.InitializeRTV(env.RTV_IrradianceSpec, mip*6+face, env.Tex_IrradianceSpec, face, mip);
-
-	// Create Irradiance Map SRVs 
-	env.SRV_IrradianceDiff = mRenderer.CreateSRV();
-	env.SRV_IrradianceSpec = mRenderer.CreateSRV();
-	env.SRV_BlurTemp = mRenderer.CreateSRV();
-	mRenderer.InitializeSRV(env.SRV_IrradianceDiff, 0, env.Tex_IrradianceDiff, false, true);
-	mRenderer.InitializeSRV(env.SRV_IrradianceSpec, 0, env.Tex_IrradianceSpec, false, true);
-	mRenderer.InitializeSRV(env.SRV_BlurTemp , 0, env.Tex_BlurTemp);
-	for (int face = 0; face < 6; ++face)
-	{
-		env.SRV_IrradianceDiffFaces[face] = mRenderer.CreateSRV();
-		mRenderer.InitializeSRV(env.SRV_IrradianceDiffFaces[face], face, env.Tex_IrradianceDiff, false, false);
-	}
-	env.SRV_IrradianceDiffBlurred = mRenderer.CreateSRV();
-	mRenderer.InitializeSRV(env.SRV_IrradianceDiffBlurred, 0, env.Tex_IrradianceDiffBlurred, false, true);
-
-	// Queue irradiance cube face rendering
-	mbRenderEnvironmentMapIrradiance.store(true);
-
-	//assert(mpScene->mIndex_ActiveEnvironmentMapPreset == static_cast<int>(ActiveEnvMapIndex)); // Only false durin initialization
-	mpScene->mIndex_ActiveEnvironmentMapPreset = static_cast<int>(ActiveEnvMapIndex);
-
-	// Update HDRMetaData when the nvironment map is loaded
-	HWND hwnd = mpWinMain->GetHWND();
-	mEventQueue_WinToVQE_Renderer.AddItem(std::make_shared<SetStaticHDRMetaDataEvent>(hwnd, this->GatherHDRMetaDataParameters(hwnd)));
-
-}
-
-void VQEngine::UnloadEnvironmentMap()
-{
-	FEnvironmentMap& env = mResources_MainWnd.EnvironmentMap;
-	if (env.Tex_HDREnvironment != INVALID_ID)
-	{
-		// GPU-sync assumed
-		mRenderer.GetWindowSwapChain(mpWinMain->GetHWND()).WaitForGPU();
-		
-		mRenderer.DestroySRV(env.SRV_HDREnvironment);
-		mRenderer.DestroySRV(env.SRV_IrradianceDiff);
-		for (int face = 0; face < 6; ++face) mRenderer.DestroySRV(env.SRV_IrradianceDiffFaces[face]);
-		mRenderer.DestroySRV(env.SRV_IrradianceSpec);
-		mRenderer.DestroySRV(env.SRV_HDREnvironmentDownsampled);
-		mRenderer.DestroySRV(env.SRV_BlurTemp);
-		mRenderer.DestroySRV(env.SRV_IrradianceDiffBlurred);
-		// mRenderer.DestroyUAV(); // TODO:?
-		mRenderer.DestroyTexture(env.Tex_HDREnvironment);
-		mRenderer.DestroyTexture(env.Tex_IrradianceDiff);
-		mRenderer.DestroyTexture(env.Tex_IrradianceSpec);
-		mRenderer.DestroyTexture(env.Tex_HDREnvironmentDownsampled);
-		mRenderer.DestroyTexture(env.Tex_IrradianceDiffBlurred);
-
-		env.SRV_HDREnvironment = env.Tex_HDREnvironment = INVALID_ID;
-		env.SRV_HDREnvironmentDownsampled = env.Tex_HDREnvironmentDownsampled = INVALID_ID;
-		env.SRV_IrradianceDiff = env.SRV_IrradianceSpec = INVALID_ID;
-		env.MaxContentLightLevel = 0;
 	}
 }
